pyproject.toml
[tool.poetry]
name = "tilly-dashboard"
version = "0.1.0"
description = ""
authors = ["eplastun <yeplastun@gmail.com>"]
readme = "README.md"
packages = [{include = "tilly_dashboard"}]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^1.5.3"
plotly = "^5.13.1"
streamlit = "^1.20.0"
psycopg2-binary = "^2.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
